resolvendo bug da selecao de grupode de limpesa do dia

+++ src/main/java/com/example/CleanList/Group/GroupRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 
 public interface GroupRepository extends JpaRepository<Group, Long> {
-    Group findByOrder(@Param("order") int order);
+    Group findByOrder(@Param("order") long order);
 }
